Define svsm-attest library with SVSM proxy input
As SVSM communication with the proxy will take place over a socket, a
standard format on how to (de)serialize data to/from the socket must be
agreed upon.

SvsmProxyInput defines this format. SVSM will only need to send data
specific for its attestation and need not concern itself with the
underlying protocol that the proxy will use to attest.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>